moved MyUtils class to new "util" package
