+++ core/src/miniventure/game/item/TileItem.java
@@ -3,7 +3,7 @@ package miniventure.game.item;
 import java.util.HashMap;
 
 import miniventure.game.GameCore;
-import miniventure.game.MyUtils;
+import miniventure.game.util.MyUtils;
 import miniventure.game.world.WorldObject;
 import miniventure.game.world.entity.mob.Player;
 import miniventure.game.world.tile.CoveredTileProperty;
