+++ core/src/miniventure/game/util/MyUtils.java
@@ -1,4 +1,4 @@
-package miniventure.game;
+package miniventure.game.util;
 
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
